+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 
 import { createTutorialStringIds } from './tools/createTutorialStringIds';
 import { getStringContents } from './tools/getStringContent';
+import { deleteStringIds } from './tools/deleteStringIds';
 
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
@@ -12,6 +13,11 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand('cnc-mission-stringids.getStringContents', () => {
 			getStringContents();
+		})
+	);
+	context.subscriptions.push(
+		vscode.commands.registerCommand('cnc-mission-stringids.deleteStringIds', () => {
+			deleteStringIds();
 		})
 	);
 }